Fix a bug in KeyBinder
mwse.loadConfig recursively sets all the missing fields in the loaded table to the fields from defaultConfig. If defaultConfig had keyCode = tes3.scanCode.i, and the myModConfig.json had mouseButton = 2, then the loaded table would end up with both. That wasn't intended. The mwseMCMKeyMouseCombo table was meant to have only one of: `mouseWheel`, `mouseButton` and `keyCode`.

This fixes it by writing false to the two unused table fields.

Adjust tes3.isKeyDown to account for this