MCM: implement internal search handler.
This searches through all the pages in a mod's MCM. Checks their labels, description and text fields. Then goes to search recursively through the `page.components` and `page.sidebar`.

The search speed is instant in an install with roughly 20 mod MCMs, so I decided to turn `searchDescriptions` to on by default. Can be changed if it proves to be slow on larger modlists.