update `DependencyManager`, & make Logger.get check `moduleName`
updated the `DependencyManager` files so that they make use of the `moduleName` field of loggers. This means they now display as

`DependencyManager (<package.name>)`
instead of
`<package.name>.DependencyManager`

i also updated the `Logger.get` function so that it will check `moduleName`s, if that parameter is provided. I think that this function should ideally accept a table parameter