Reduce usage of `getComponent` in `SideBarPage.lua`
In three different parts of the `SideBarPage.lua` file, the `Component:getComponent` function was used to create a new instance of an already known class.

Each of those calls to `getComponent(self, data)` would ultimately result in the following lines of code being executed:
```lua
data.parentComponent = self

... -- complicated logic to fetch `componentClass`

return componentClass:new(data)
```

Since `componentClass` is already known ahead of time, it's possible to replace `getComponent(self, data)` with the `componentClass:new(data)`.

This makes the code easier to follow, and (ignoring the three import statements that were added), it does not result in a net increase in lines of code.