Add mwseMCMColorPickerButton.
This setting has a color preview showing the currently selected color and a button. Clicking on the button will open a new menu with a color picker. This is meant for SideBar and Filter MCM pages, since one the left part of the page might be too narrow for big inline color picker widgets. This compact setting should fit in any page.