rename `Logger` to `mwseLogger`
Also rebuilt documentation, and removed some of the documentation from the actual Logger implementation file.

The changes to `activate.md` are a result of rebuilding the documentation after undoing a previous commit where I accidentally changed the autocomplete file.

Not all of the documentation for the logger has been finalized yet. I'm still waiting to see how the guides will fit into all of this.